requirements.txt
Django==4.0.5
psycopg2==2.9.3
djangorestframework==3.13.1
django-filter==22.1
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.11.0
django-apscheduler>=0.6.2
requests>=2.27.1
configparser>=5.2.0
drf-writable-nested>=0.6.3
django-request-logging>=0.7.5
python-json-logger>=0.1.11
python-dateutil>=2.8.2
firebase-admin>=5.2.0
APScheduler~=3.9.1
google-cloud-vision==3.1.2
language~=0.6
python-dateutil==2.8.2
channels==3.0.5
channels_redis==3.4.1
git+https://github.com/potatojoayo/graphene-django.git@main#egg=graphene-django
django-graphql-jwt==0.3.4
graphene~=3.1.1
celery>=5.2
Redis>=4.3.4
Pillow>=9.2.0
django-tailwind==3.4.0
django-browser-reload==1.6.0
django-model-utils==4.3.1